Fixed full page spinner for interactive sessions actions after Turbo Stream update
