+++ apps/dashboard/app/helpers/batch_connect/sessions_helper.rb
@@ -96,7 +96,7 @@ module BatchConnect::SessionsHelper
     button_to(
       batch_connect_session_contexts_path(token: batch_connect_app.token),
       method: :post,
-      class: %w[btn px-1 py-0 btn-outline-dark relaunch],
+      class: %w[btn px-1 py-0 btn-outline-dark relaunch full-page-spinner],
       form_class: %w[d-inline relaunch],
       title: title,
       'aria-label': title,
@@ -112,7 +112,7 @@ module BatchConnect::SessionsHelper
     button_to(
       new_batch_connect_session_context_path(token: session.token),
       method: :get,
-      class: %w[btn px-1 py-0 btn-outline-dark],
+      class: %w[btn px-1 py-0 btn-outline-dark full-page-spinner],
       form_class: %w[d-inline edit-session],
       title: title,
       'aria-label': title,
