+++ apps/dashboard/app/javascript/batch_connect_settings.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import { bindFullPageSpinnerEvent } from './utils';
+
+jQuery(function (){
+  bindFullPageSpinnerEvent();
+});
