+++ apps/dashboard/app/javascript/turbo_shim.js
@@ -20,14 +20,19 @@ export function replaceHTML(id, html) {
   }
 }
 
-export function pollAndReplace(url, delay, id) {
+export function pollAndReplace(url, delay, id, callback) {
   fetch(url, { headers: { Accept: "text/vnd.turbo-stream.html" } })
     .then(response => response.ok ? Promise.resolve(response) : Promise.reject(response.text()))
     .then((r) => r.text())
     .then((html) => replaceHTML(id, html))
-    .then(setTimeout(pollAndReplace, delay, url, delay, id))
+    .then(() => {
+      setTimeout(pollAndReplace, delay, url, delay, id, callback);
+      if (typeof callback == 'function') {
+        callback();
+      }
+    })
     .catch((err) => {
-      console.log('Cannot retrive partial due to error:');
+      console.log('Cannot retrieve partial due to error:');
       console.log(err);
     });
 }
