+++ apps/dashboard/app/views/batch_connect/settings/show.html.erb
@@ -1,5 +1,9 @@
 <% content_for :title, t('dashboard.bc_saved_settings.title') %>
 
+<%- content_for :head do -%>
+  <%= javascript_include_tag('batch_connect_settings', nonce: true) %>
+<%- end -%>
+
 <%= render partial: 'batch_connect/shared/breadcrumb',
 locals: {
   links: [
