+++ apps/dashboard/app/views/batch_connect/shared/_full_page_spinner.html.erb
@@ -1,6 +1,3 @@
-<%- content_for :head do -%>
-  <%= javascript_include_tag('full_page_spinner', nonce: true) %>
-<%- end -%>
 <div id="full-page-spinner" class="d-none">
   <div class="spinner-border" role="status"></div>
 </div>
